+++ Options/CHANGELOG.lua
@@ -5,6 +5,7 @@ TMW.CHANGELOG_LASTVER="7.4.0"
 TMW.CHANGELOG = [==[
 ## v11.0.1
 * Fix: #2174 - Autocast conditions not working
+* Fix: #2176, #2177, #2175, Many spells not tracking cooldowns properly
 
 ## v11.0.0
 * Added the original Lightwell icon type back to Cata Classic